Design study: 32-bit integers and 128-bit Apps!
This describes how we achieve two seamingly mutually exclusive goals:
supporting full 32-bit primitive numbers and fitting heap applications
in 128-bits.

Atom representation

All atoms are represented as 33-bit values, where the top bit
indicates whether the rest is an unboxed 32-bit primitive integer or a
tagged value.  Tagged values have bottom HeadTag (HT) bits reserved.

Effectively we have a multi level tagging scheme (assume HT=7 marked
with -------):

- Integers            1NNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNN integer
- Tagged Values
  = Pointers          01sppppppppppppppppppppppp------- shared,heap index
  = Non-pointers
    * Constructors    00000.AAAiiiiiiiiiiiiiiiii------- arity,index
    * Primitives      00001sAAAiiiiiiiiiiiiiiiii------- arity,swapped,index
    * Arguments       00010s...iiiiiiiiiiiiiiiii------- shared,index
    * Registers       00011s...iiiiiiiiiiiiiiiii------- shared,index
    * Functions       00100oAAAiiiiiiiiiiiiiiiii------- original,arity,index
    * Invalid         00101lr..iiiiiiiiiiiiiiiii------- LUT, REGID, index

Recall, Argument and Register atoms can only occur in templates.
Invalid is used to mark unused atom slots and implicitly represents
the size.  Furthermore, for CASE/PRIM, the LUT/RegId is encoded in an
Invalid atom in the last atom slot with the corresponding LUT/REGID
bits set.  For simplicity most atoms share layout, but that could be
change if some, say functions, needs a larger address space.

Heap applications

While we are primarily concerned with heap allocated apps as opposed
to apps in templates, which could have a completely different
representation with hardly any consequence, we sacrifice pointer
address range for a simpler representation.

The HT bits in the first atom are the application tag bits.  For the
case where the first atom is a primitive number, we recover the HT
bits from the second atom.  This is possible because it is (currently)
impossible for a legal heap allocation to have both the first and the
second atom being a number. (Heap app can't be PRIM and can't contain
REG and ARG atoms which could further enables us to play a few more
tricks to save bits here and there).

These are the heap tag bits:
- HT_INT0, HT_INT1, HT_INT2, HT_INT3: which atom are numbers
- HT_GC: COLLECTED or {AP, CASE, PRIM}
- HT_NF: application is an irreducible normal form

Thus, HT is currently 6, however it could be reduced to 4 as follows:

- HT_NF is strictly redundant and can be recovered, but is included
  currently as in the original Reduceron-2 as the payoff is high
  compared to the cost.

- Only primitive applications can have a number at the head, so it may
  be possible to represent them swapped thus the need for NUM_TAG0
  disappears.

Comparing the App tags from the original emulator:

- The size isn't explicitly represented, but instead unused atoms are
  marked with an illegal value.

- lut and regId take the place of the last atom (apparently the
  compiler already expects this for LUTs and PRIMs are all currently
  binary).

- The App tag is implicitly represented by a combination of the GC
  flag and the last atom.

With an HT of 6, we left with 30 - 6 = 24 bits for the pointer, enough
for 16 Mi heap cells, a 256 MiB heap.  The more complicated scheme
reduces HT to 4 bit, thus leaving 26 bits for the pointer (= 64 Mi
cells = 1 GiB).

This tagging scheme comes from the 128-bit constraint of external
memory.  Rather than operating on the packed Atoms and Apps directly,
it's probably better to only keep Apps packed in the main heap and
just expand Reducerons current representation to 35-bit Atoms and
145-bit Apps everwhere else, notably also in the cache, even though it
comes at a slight memory overhead.