Reduceron in less than 1 Mb! Runs on DE2-70
Reduced the heap to a quarter size, the code store to half the size,
and shrunk all other memories as much as the emulation run of the
benchmark suggested would work.

Note, this is only Step 2 in the master plan, and as such is a
temporary step, however it enables me to test the progress made on
cheaper development kits.

The shunk code size means KnuthBendix and Mate can't fit.  Of the
remaining 17 benchmarks, 13 work correctly (see
Reduceron-reduced.txt).  The four failures are probably expected given
the heap and integer restructions.