Flite: do not use INT 0 for constructors without destructors
The value is never deconstructed which suggests we should even
generate anything for it in first place, but this is a minimal change
that enables the property "No Valid Reduceron App has Two Consecutive
Primitive Integers" which the 32/128 encoding scheme depends on.

I use CON 0 66 to make it stand out.