ld32, first baby steps
Deciding on the model for this took a long time, but in the end, this
seems the simplest possible implementation which also appear to be the
natural direction for Reduceron.

ld32 quite simply is a binary primitive that loads from the address
given as the first argument, ignores it's second argument, and return
the loaded value.

To get the sequential behavior we depend on the behavior so far only
implemented in the emulator:

  N F -> F N, where F is not a prim

This isn't expected to be too hard to implement in Lava Reduceron.